Fixed bugs and added RequiredPermissions

+++ src/main/java/me/dedose/commandhandler/api/CommandExecutor.java
@@ -1,5 +1,6 @@
 package me.dedose.commandhandler.api;
 
+import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.events.message.guild.GuildMessageReceivedEvent;
 import net.dv8tion.jda.core.hooks.ListenerAdapter;
 
@@ -12,24 +13,45 @@ public class CommandExecutor extends ListenerAdapter {
         CommandHandler cmdHandler = new CommandHandler();
         String command = event.getMessage().getContentRaw().split(" ")[0];
         String[] args = Arrays.copyOfRange(event.getMessage().getContentRaw().split(" "), 1, event.getMessage().getContentRaw().split(" ").length);
-        System.out.println(cmdHandler.commands);
+
         for(Command cmd : cmdHandler.getCommands()){
             for(String aliases : cmd.getAliases()){
                 System.out.println(aliases);
-                if(command.equalsIgnoreCase(cmdHandler.getPrefix() + aliases)){
-                    System.out.println(args.length);
-                    if(args.length >= cmd.getRequiredArgs()){
-                        if(cmd.getUsableChannels() == null) {
-                            cmd.getCommandExecuteListener().onCommand(event.getMember(), event.getChannel(), args, event.getMessage());
-                        }else{
-                            if(cmd.getUsableChannels().contains(event.getChannel())){
+                if(command.equalsIgnoreCase(cmdHandler.getPrefix() + aliases)) {
+                    if (cmd.getRequiredPermissions().isEmpty()) {
+                        if (args.length >= cmd.getRequiredArgs()) {
+                            if (cmd.getUsableChannels().isEmpty()) {
                                 cmd.getCommandExecuteListener().onCommand(event.getMember(), event.getChannel(), args, event.getMessage());
-                            }else{
-                                event.getChannel().sendMessage(cmdHandler.getIncorrectChannelMessage()).complete().delete().queueAfter(15, TimeUnit.SECONDS);
+                            } else {
+                                if (cmd.getUsableChannels().contains(event.getChannel())) {
+                                    cmd.getCommandExecuteListener().onCommand(event.getMember(), event.getChannel(), args, event.getMessage());
+                                } else {
+                                    event.getChannel().sendMessage(cmdHandler.getIncorrectChannelMessage()).complete().delete().queueAfter(15, TimeUnit.SECONDS);
+                                }
                             }
+                        } else {
+                            event.getChannel().sendMessage(cmdHandler.getIncorrectUsageEmbed().getIncorrectEmbed(cmd.getUsage(), event.getMember()).build()).complete().delete().queueAfter(15, TimeUnit.SECONDS);
                         }
                     }else{
-                        event.getChannel().sendMessage(cmdHandler.getIncorrectUsageEmbed().getIncorrectEmbed(cmd.getUsage(), event.getMember()).build()).complete().delete().queueAfter(15, TimeUnit.SECONDS);
+                        for(Permission perms : cmd.getRequiredPermissions()){
+                            if(event.getMember().getPermissions().contains(perms)){
+                                if (args.length >= cmd.getRequiredArgs()) {
+                                    if (cmd.getUsableChannels().isEmpty()) {
+                                        cmd.getCommandExecuteListener().onCommand(event.getMember(), event.getChannel(), args, event.getMessage());
+                                    } else {
+                                        if (cmd.getUsableChannels().contains(event.getChannel())) {
+                                            cmd.getCommandExecuteListener().onCommand(event.getMember(), event.getChannel(), args, event.getMessage());
+                                        } else {
+                                            event.getChannel().sendMessage(cmdHandler.getIncorrectChannelMessage()).complete().delete().queueAfter(15, TimeUnit.SECONDS);
+                                        }
+                                    }
+                                } else {
+                                    event.getChannel().sendMessage(cmdHandler.getIncorrectUsageEmbed().getIncorrectEmbed(cmd.getUsage(), event.getMember()).build()).complete().delete().queueAfter(15, TimeUnit.SECONDS);
+                                }
+                            }else{
+
+                            }
+                        }
                     }
                 }
             }
